requirements_dev.txt
# Generic python dependencies
black==24.4.2
mypy==1.10.0
pip==23.1.2
pre-commit==3.6.0
ruff==0.4.4

chromadb==0.5.0
google-generativeai>=0.8.1
ical>=8.2.0

# Home Assistant testing dependencies
pytest-homeassistant-custom-component==0.13.162
syrupy>=4.6.1

# Component dependencies
hassil>=1.7.0
home-assistant-intents>=2024.4.3

mashumaro>=3.13.1
ical>=8.1.1
